Add Assets and saves folder.
Assets will contain the json files with info. Feel free to change the
values.
SeasonPretty is the gson generated version of SeasonStart.
The saves folder contains the savegames.
Modified SeasonStart to make it readable for gson:
Driver is the only attribute of season that is a List.
Removed the list wrapper for all the other attributes. EG: Mechanic,
strategist ...
Engine is a class and not a string! The values I added are bogus, so
please take a look at them.
Add tests to SeasonTest to test the new added methods.